+++ src/main/java/com/ibm/southside4u/service/IProductService.java
@@ -1,0 +1,16 @@
+package com.ibm.southside4u.service;
+
+public interface IProductService {
+
+	
+	public void getAllProducts() ;
+
+	
+	public void getProductsAttrByName(String name);
+
+	
+	public void getProductsAttrById(Integer id);
+
+	
+	public void getProductsAttrByType(String type);
+}
