--- src/main/java/com/ibm/soutside4u/service/IProductService.java
@@ -1,16 +1,0 @@
-package com.ibm.soutside4u.service;
-
-public interface IProductService {
-
-	
-	public void getAllProducts() ;
-
-	
-	public void getProductsAttrByName(String name);
-
-	
-	public void getProductsAttrById(Integer id);
-
-	
-	public void getProductsAttrByType(String type);
-}
