--- src/main/java/com/ibm/soutside4u/service/impl/ProductServiceImpl.java
@@ -1,33 +1,0 @@
-package com.ibm.soutside4u.service.impl;
-
-import org.springframework.beans.factory.annotation.Autowired;
-
-import com.ibm.soutside4u.service.IProductService;
-import com.ibm.soutside4u.utils.CustomLogger;
-
-public class ProductServiceImpl implements IProductService {
-
-@Autowired
-CustomLogger logger;
-	@Override
-	public void getAllProducts() {
-		String methodName = "getAllProducts";
-		logger.methodEntryLogger(ProductServiceImpl.class, methodName);
-	}
-
-	@Override
-	public void getProductsAttrByName(String name) {
-
-	}
-
-	@Override
-	public void getProductsAttrById(Integer id) {
-
-	}
-
-	@Override
-	public void getProductsAttrByType(String type) {
-
-	}
-
-}
